Replace test double with sinon
There were some cases where testdouble’s api truly was nicer —
especially for verifying. But, its insistence that all stubs’ return
values be set up through the declarative `when()` interface, rather
than calling through to the underlying function, means that the
`td.when`s have to be set up again each time you create a new double
between tests (which you need to do to have an accurate count of the
invocations). Normally, that might not be so bad, except that here it
means that we either have to recreate all the states in beforeEach too
(not just the doubles), which requires those ugly outer variables to
save the created states for our tests, or we have to separate the
states from the stubs in the code — which is harder to read. Moreover,
no matter what we do, we have to add basically empty
transitionOut/twimlFor methods to keep typescript happy, so there’s
duplication.