Replace config.staticFiles.holdMusic.loopCount with a twimlFor option
This makes the holdMusic twiml more configurable while removing the
obscure loop count setting from the api (which you’d really only want
to change if your twiml was doing something custom anyway, like
alternating between Play and Say tags). Simultaneously, it replaces the
custom code in the hold music endpoint handler that was producing the
url for the hold audio file with a use of the standard urlFor function.
This gets us one step closer to eventually supporting cache busting
strategies in urlFor beyond the `?v=` query parameter, which seems to
work with twilio but is generally not a best practice (because not all
caches treat the same base uri with different query parameters as a
distinct cache key).