Static files: fix expiration time
Before, static-expiry was using a one year expiration time just by
default, while we were passing a two year expiration time to
express.static. But then express.static was rounding that two years
down to one behind the scenes, which meant that it just happened to
match the one year being used by static-expiry. And that was good,
because our test for the cache control header didn’t include the
version parameter, which meant static-expiry ignored the request and
actually wasn’t adding the cache-control header at all, and instead
that was coming from express.static. So, basically, things seemed to be
working (except for the 2y turning into one), but everything was
actually fucked up.

Now, we explicitly pass the same value (one year) to static-expiry and
express.static. We pass a value (one year) that express.static won’t
round down. And we make sure static-expiry’s header adding facilities
are being invoked in our tests, by passing in the versioned url and
checking for the expires header too.

Meanwhile, there was actually a bug in static-expiry too when it comes
to adding headers, but that bug only applied to connect — so it just
confused the situation. Regardless it’s now fixed.