Support customizing static file handling (url fingerprinting and serving behavior)
This commit does a bunch of things:

- The user can now pass in their own function for fingerprinting static
file urls, and this function will be passed to the higher-order
function that makes the `urlFor` function that the states ultimately
see. (Also, that higher order function has now been renamed
`makeUrlFor` rather than `urlFor`—yay clarity!) This allows other
fingerprinting strategies to be used in more advanced deployment
scenarios.

- As part of the above change, `makeUrlFor` was also cleaned up.
Before, when it was fingerprinting a url, it would manually strip the
mount path (since static-expiry’s furl doesn’t understand mount paths)
and then add it back in at the end. That was an inappropriate coupling,
though, between the larger urlFor function and (the quirks of)
static-expiry’s furl. This was clearly unsustainable now that the
fingerprinting function is customizable, as the user’s provided
function may be mount-path aware. So, instead, that functionality
around handling mount paths is moved explicitly to a static-expiry
helper function, which takes the underlying furl and wraps it with the
mount path stripping + re-adding ability.

- When the user provides their own fingerprinting function, they must
also provide their own middleware to reads those fingerprinted urls and
returns the underlying files. This is done using the middleware option
that existed before, though that now has a broader meaning of: go from
fingerprinted url to HTTP response (with contents and headers), whereas
before it just meant add contents (not headers) from the
already-fingerprint-removed url. The user is allowed, but not required,
to provide a custom middleware function even if they’re using the
built-in fingerprinting functionality.

- The hold music endpoint now lives under the static files mount path,
if any. Especially now that the user can provide their own middleware
for resolving fingerprinted urls (of which the hold music endpoint is
one), mounting all those middleware under a single path and having just
one chain that handles serving for both static files and the hold music
endpoint seemed easier to reason about.

- Before, we used private static-expiry implementation details to add
records to its caches so that `furl(holdMusicEndpoint)` would contain
the version parameter of the hold music mp3 file. This wasn’t really
appropriate, though, because not only did it use private internals, but
static-expiry also couldn’t actually go the other way (i.e., from the
holdMusic endpoint to a static file—of course), so the cache entries
were misleading. Now, rather than modifying the cache to change furl’s
behavior directly, we just wrap furl in a function that checks if the
path is the hold music endpoint and then versions that alone (looking
up the version of the underlying file in static-expiry’s cache…which is
still private, but less janky and with no side effects).

- In the holdMusic endpoint’s twimlFor, we use `path.normalize` rather
than `path.join` to connect the mountPath and the holdMusicPath, as
`path.join` will give the wrong separator for a url on Windows.
`path.normalize` also allows our holdMusicPath to be a bit more
flexible (e.g. “fileName” and “./fileName” are both supported now).
This flexibility is mirrored in the staticExpiryHelpers’ `path.resolve`
use to calculate the cache key in `getFingerprintForFile`.