Static Expiry Helpers: more defensive
Context: mountPathAwareFurl is only used when the built-in,
static-expiry-based furl is being used. Then, it wraps that furl to
account for the mountPath.

This change does two things:

1.) It normalizes the case where mountPath === ‘/‘. See comment for the
backstory here.

2.) Throws an error if the path to fingerprint points to something
that’s not under the mount path. That was previously not an error
because there was one case—the hold music endpoint—where we were
hacking static-expiry’s cache such that we could be fingerprinting a
url not under the mount path. Now, though, there are no valid such
cases.

3.) We handle this error specially in urlFor