Make middleware option optional in all cases
Conceptually, generating fingerprinted urls and actually serving the
files that those urls correspond to are separate tasks. Twilio-ivr
genuinely does need to know how to do the fingerprinting part, so that
the rendered twiml can include the fingerprinted url. However, it
doesn’t need to know how to serve the static files. In practice, it can
be convenient for twilio-ivr to serve the static files, as that’s
simpler than setting up another way to do it. But, *requiring* that
twilio-ivr do the file serving is clearly a mistake — it just won’t
work for many setups (e.g., what if the files are on a cdn?) and it’s
conceptually an unnecessary coupling of concerns.

Requiring the `staticFiles.middleware` option be set when the
`staticFiles.fingerprintUrl` option was used suggested that the library
required that it always know how to actually serve the files. By making
this setting optional, we make clear that that is not the case.

Of course, in practice, the user could always serve their static files
without going through twilio-ivr even when the middleware option was
required — they could just provide twilio-ivr a no-op function, and
have their static file requests never reach twilio-ivr (using an
earlier middleware or even a reverse proxy sitting in front of node) —
but this has been bothering me conceptually for a while so it’s nice to
fix it.