Breaking: Fix Host-header security bug; bump version
Here, we require that the user provide (a function returning) a string
to use as the host part in generated absolute urls. This is instead of
the library getting that host by reading `req.headers.host`. Reading
the Host header was a security vulnerability because the requester can
actually put whatever they want in that field. This security
vulnerability probably wasn’t exploitable in practice, as long as the
`validate` was left on, because then only requests from Twilio would be
accepted (and twilio can be trusted to provide the correct Host
header). Still…

Another theoretical shortcoming with reading the Host header—though,
again, not applicable to how Twilio would have sent requests in
practice—is that it can simply be missing, which would’ve caused things
to blow up. For both these reasons, the new approach is much better.

While we’re at it, this commit also makes it possible for the caller of
urlFor to override the host/scheme on a per-call basis, which is a nice
way (e.g.) to handle static assets that are served from multiple hosts
for load-balancing or whatever (probably not necessary in a phone tree,
but can’t hurt).